Move the container block to the root, and register the child blocks via the child block functionality built into the Jetpack register block function.
