+++ class.jetpack-plan.php
@@ -63,10 +63,6 @@ class Jetpack_Plan {
 				'akismet',
 				'recurring-payments',
 				'premium-content/container',
-				'premium-content/buttons',
-				'premium-content/logged-out-view',
-				'premium-content/subscriber-view',
-				'premium-content/login-button',
 			),
 		),
 		'premium'  => array(
