+++ extensions/blocks/premium-content/_inc/blocks.js
