+++ extensions/blocks/premium-content/_inc/controls.js
