+++ extensions/blocks/premium-content/_inc/icons.js
