+++ extensions/blocks/premium-content/_inc/inspector.js
