+++ extensions/blocks/premium-content/_inc/new-plan.js
