+++ extensions/blocks/premium-content/_inc/premium.js
