+++ extensions/blocks/premium-content/_inc/tab.js
