+++ extensions/blocks/premium-content/_inc/tabs.js
