+++ extensions/blocks/premium-content/editor.scss
