+++ extensions/blocks/premium-content/login-button/index.js
@@ -58,6 +58,7 @@ const settings = {
 	parent: [ 'jetpack/premium-content/container' ],
 	keywords: [ __( 'link', 'jetpack' ) ],
 	supports: {
+		inserter: false,
 		align: true,
 		alignWide: false,
 		html: false,
