--- extensions/blocks/premium-content/subscriber-view/editor.js
@@ -1,8 +1,0 @@
-/**
- * Internal dependencies
- */
-import registerJetpackBlock from '../../../shared/register-jetpack-block';
-import { name, settings } from '.';
-
-const noPrefix = true;
-registerJetpackBlock( name, settings, [], noPrefix );
