+++ extensions/blocks/premium-content/subscriber-view/index.js
@@ -13,21 +13,18 @@ import { __ } from '@wordpress/i18n';
 
 const name = 'premium-content/subscriber-view';
 const settings = {
-	attributes: {},
-	/* translators: block name */
 	title: __( 'Subscriber View', 'jetpack' ),
-	/* translators: block description */
 	description: __( 'Subscriber View.', 'jetpack' ),
+	icon,
 	category: 'grow',
+	attributes: {},
+	edit,
+	save,
 	parent: [ 'premium-content/container' ],
 	supports: {
-		// Hide this block from the inserter.
 		inserter: false,
 		html: false,
 	},
-	edit,
-	icon,
-	save,
 	deprecated,
 };
