+++ extensions/blocks/premium-content/view.js
