+++ extensions/blocks/premium-content/view.scss
