+++ extensions/index.json
@@ -41,11 +41,7 @@
 	"beta": [
 		"amazon",
 		"anchor-fm",
-		"premium-content/container",
-		"premium-content/buttons",
-		"premium-content/logged-out-view",
-		"premium-content/login-button",
-		"premium-content/subscriber-view"
+		"premium-content"
 	],
 	"experimental": [ "seo" ],
 	"no-post-editor": [
